Test a user can be destroyed
Tests a user can be destroyed. Creates a new role and user and assigns
the relationship between them.

Calls the destroy method and passes the User ID. Asserts that the
database is missing a user with the ID, role_id and email address
reflected in the test in the users table.