Allow only admins to create new schools
Creates a new school policy method to check whether the user has the
role 'Administrator'. If the user is authenticated as an administrator
they can post to the schools.store route and create new schools. If not,
they should be redirected with an error message. Tests whether an admin
user can create new schools.