Allow authenticated user to update schools
Configures the SchoolController update method to allow schools to be
updated where the authenticated user can update the school. Creates an
update method on the SchoolPolicy class to check for authentication.
Tests that an administrator can update existing schools. Sends a PUT
request to the schools/{school} endpoint passing updated data. Asserts
the database has a record with the updated ID. Asserts the user is
redirected and the session has an alert message.