Test authentication to school edit view
Tests the authentication to the schools edit view to edit an existing
school. Asserts that only an administrator can access the view. Asserts
that the correct school exists on the view. Asserts that non admin users
cannot access the view and asserts that they are redirected with an
error message in session.