Authenticate access to viewing assets
Authenticates access to viewing assets by checking that the
authenticated user can access the school that the asset is associated
with. Tests that a user can only view assets if associated with the
given school. Asserts that the user can see the show view for the asset
or is redirected with an error message in session if unauthorised.