Allow authenticated users to create assets
Allows users who are associated with a particular school to update that
schools assets if they have the correct roles.

Tests that users associated with a school and appropriate role can update existing
assets.

Tests users associated with a school but without the appropriate
role cannot update existing assets.

Asserts users without permission are redirected with an error message in session.

Asserts users with permission have updated the asset in storage and are redirected to assets.show view with success message in session.