Handle failed authorization attempts on categories
Handles any failed authorization attempt on assets by throwing a new
UnauthorizedException. Checks if current user can either 'create' or
'update' assets respectively in the authorize() method of the form
request.

Overrides the failedAuthorization() method on the extended FormRequest
class to throw a new exception when the user is unauthorized to perform
the given action.

Updates all tests to reflect new updates to authorization.