Add errors blade file and style schools.create
Adds a new errors partial to be included where any errors can be
displayed. Updates the settings view to link to the schools.create view
and formats the view to use appropriate styling classes. Adds the errors
partial to the schools.create view to be displayed if any errors are
thrown. Adds client side validation to the schools.create view.