Create reactive type selector component
Creates a category selector component and a type selector component.
Types returns all of the asset types from the database. Categories
returns the selected types categories. If a user changes the asset type,
the categories for that specific asset type are returned.

To do this the types component emits an event on change and sends the
current type ID. The categories component listens for the event and
updates the type ID in it's own component. It then performs the axios
request again to return the new categories for the selected asset type.