Move type parameter into axios params
Moves the form parameter from the end of the GET URL request and into
the params object to keep it tidy. Also re-orders the requests in the
API file to be more consistent with how the application is ordered i.e.
select a school, select a type, select a category.