Add asset fields to create asset form
Adds necessary fields to asset form to create new assets. Adds v-if
check to check whether the selected asset type ID is the ID matching IT
Equipment; if so, the additional fields show. Adds listener on the main
Vue root instance to listen for typeChanged event and update the root
type ID accordingly.