Add ignore method to unique rule for update requests
Adds an ignore() method to the unique rule when checking the asset tag
on update. Checks if the asset tag provided by the user is unique in the
assets table where the school ID matches the provided school, ignoring
the asset that is being updated. That way no error will be thrown if the
asset is updated but the tag is the same.