Refactor store method on AssetController
Moves the logic for creating the asset and finance record out of the
controller and into the StoreAsset form request class under a persist()
method for clarity and condenseness in the controller. Calls the persist
method in the store method. Updates the if check on storing an asset to
check if $asset->wasRecentlyCreated instead of using exists.