Save finance in pence and return in pounds
Creates mutators on the Finance model to save the finance values to the
database in their lowest form; in pence. Creates accessors on the
Finance model to return the finance values from the database in pounds.
Tests that the accessors and mutators work correctly. Updates the
finances migration file to save the finance values as big integer
values.