Add accessors to format dates on Finance model
Adds multiple new accessors to format the dates on the Finance model.
Returns the dates without the time in Y-m-d format. Creates an accessor
for returning the accounting start year. Creates an accessor for
returning the accounting end year. Creates an accessor for returning the
accounting start and end years combined. Tests all the new accessors
return the correct data in the correct format.