Eager load relationships for school assets
Creates a method on School model to eager load the schools assets with
the category, type and finances relationships for cleanliness in
controller. Calls the method in the show method on the
SchoolAssetController and gets the collection to return to the
view. Displays the category and type on the view.